feat: Sprint 57 — style_id u8 narrowing (P2) + is_heading_terminator tests (P3)
S57-P2: narrow HwpParagraph.style_id from u16 → u8.
HWP 5.0 spec: nStyleID is UINT8, max value 255. The u16 field type was a
historical artifact from before Sprint 56's PARA_HEADER read fix. Narrowing
removes the foot-gun: a reader of model.rs could assume 65535 styles are
addressable, which is spec-incorrect and never happens in practice.
- model.rs: `pub style_id: u16` → `pub style_id: u8`
- dispatcher.rs: `u16::from(rec.data[6])` → `rec.data[6]` (no conversion needed)
- All other call sites (`style_id: 0` literals, `as usize` casts) unchanged.

S57-P3: add is_heading_terminator policy tests.
Pins the character classification rules so future edits cannot silently flip
allowed↔blocked without failing a test.
- Canonical allowed: space, \t, (, ), 「, 」, （, ）, ., :, ·, ㆍ, …
- Canonical blocked: Korean particles 은/에/이/가/의, ASCII letter, ASCII digit

1206 unit tests + 11 fixture tests (0 ignored, 0 failed).

+++ src/hwp/control/dispatcher.rs
@@ -44,7 +44,7 @@ pub(crate) fn extract_paragraphs_from_range(
                 // nStyleID is UINT8 at byte[6]; reading two bytes would corrupt the
                 // value with the flags byte when any flag bit is set.
                 let style_id = if rec.data.len() >= 7 {
-                    u16::from(rec.data[6])
+                    rec.data[6]
                 } else {
                     0
                 };
